Update the rules engine to thread an env map. Key Nanopass ideas implemented.
Env map allows arbitrary information to be threaded through as the rules
are matched. Updates to the env can be captured so sophisticated
operations that are usually done with mutability can be done with pure
functional code.

Add functionality to >- mode marker in the directed rule combinator. It
allows subexpressions to be passed to an external function before being
given to the handler. Combines with the -> and $ modes.

Fix sequence capture with optionals and with non-sequential global captures.

Non-macro rules, remove the arity-3 rule. Add a little bit of macro
magic to bind values from the env to locals in the rule handler.

Various minor metadata improvements

Nanopass features implemented:

- define and derive dialects (languages in Nanopass)
- defpass and let-rulefn together do most of what the Nanopass defpass
  does, but with very minimal macros.
  * defpass mostly just does a bit of wrapping and a little code
    shuffling to achieve that.
  * let-rulefn shuffles some pieces around to mimic letrec, and builds
    the on-mutual call which is a bit complex to build on its own from
    the basic pieces provided