Replace use of substitute with sub in rules without a handler body.
This should run a lot faster since sub is precompiled. This seems mostly
good now. There are a few trivial test fixes to come in the next commit,
but something in some of the more elaborate dialects is broken. Magic
test suite is already green.

I think the problem is related to the use of ?:letrec. Does sub support
that?