The on-match event reverted the env to a previous version in a specific case
If a rule inside a directed rule is itself a combinator, and that
combinator executes multiple matches where changes to the matcher env
are important, subsequent matches would break because the env would be
reverted by the event handler to the version last seen by the outer
directed combinator.

I've run into this issue multiple times in the past but never could find
the cause since the fact that an event handler is injected into the flow
is not obvious while debugging the inner combinator where the issue
appeared to originate.