Playing with the idea of rebuilding rules, but currently stuck on pattern-args
Pattern-args together with rule-fn-body bind the names in the pattern to
bindings in a function wrapper for the handler body. Because the pattern
is not literally present in rebuild-rule, there isn't an obvious way to
recreate that wrapper with a new body without providing the literal
pattern again.

I could try to analyze the provided function for free variables or
provide a macro for the user to specify variables. The first seems like
too much work and the second seems to defeat the idea of doing automated stuff.